Major project structure update
* Move all packages in yi.* to
  codes.nibby.yi.* namespace

* Rename main project 'yi.editor' to 'app'

* Rename domain models project 'yi.core' to 'models'

* Merge all component modules into main app module.
  Classes in the 'share' module have had its package
  re-organized.

* Rename classes with 'Editor' prefix into 'App'

The original fragmented project structure originally
began as a curiosity experiment to learn about Java 9
module system. Furthermore, the original vision
for Yi is to create a set of reusable components to
build a collection of tools from. Now a year later,
this vision is changing. Having other time commitments
mean Yi will see far less development effort to turn
it into the swiss army knife for Go players.

As a result, these changes mainly redact the vision
of the project. All the modules are collated into
the editor project, with the main feature focus being
AI analysis once the maintenance changes are complete.